requirements.txt
beautifulsoup4==4.11.2
bs4==0.0.1
cachelib==0.10.2
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-Admin==1.6.0
Flask-Login==0.6.2
Flask-Session==0.4.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
oauthlib==3.2.2
python-dotenv==0.21.1
soupsieve==2.3.2.post1
Werkzeug==2.2.2
WTForms==3.0.1
